+++ pulse_lib/acquisition/measurement_converter.py
@@ -6,6 +6,8 @@ from pulse_lib.segments.segment_measurements import measurement_acquisition, mea
 from pulse_lib.acquisition.iq_modes import iq_mode2func
 from qcodes import MultiParameter
 
+logger = logging.getLogger(__name__)
+
 
 @dataclass
 class SetpointsSingle:
@@ -320,7 +322,7 @@ class MeasurementConverter:
         if total_selected > 0:
             self._values = [np.sum(result*accepted_mask)/total_selected for result in values_unfiltered]
         else:
-            logging.warning('No shot is accepted')
+            logger.warning('No shot is accepted')
             self._values = [np.full(len(values_unfiltered), np.nan)]
 
     def set_channel_data(self, data):