+++ pulse_lib/keysight/qs_uploader.py
@@ -14,6 +14,7 @@ from pulse_lib.tests.mock_m3202a_qs import AwgInstruction, AwgConditionalInstruc
 from pulse_lib.tests.mock_m3102a_qs import DigitizerInstruction
 from pulse_lib.segments.utility.rounding import iround
 
+logger = logging.getLogger(__name__)
 
 class AwgConfig:
     MAX_AMPLITUDE = 1500 # mV
@@ -75,7 +76,7 @@ class QsUploader:
         awg = list(self.AWGs.values())[0]
         return awg.convert_prescaler_to_sample_rate(awg.convert_sample_rate_to_prescaler(sample_rate))
 
-    def get_num_samples(self, acquisition_channel, t_measure, sample_rate):
+    def actual_acquisition_points(self, acquisition_channel, t_measure, sample_rate):
         raise NotImplementedError()
 
     def get_roundtrip_latency(self):
@@ -124,13 +125,13 @@ class QsUploader:
         aggregator.upload_job(job, self.__upload_to_awg) # @@@ TODO split generation and upload
 
         duration = time.perf_counter() - start
-        logging.debug(f'generated upload data ({duration*1000:6.3f} ms)')
+        logger.debug(f'generated upload data ({duration*1000:6.3f} ms)')
 
     def __upload_to_awg(self, channel_name, waveform):
 #        vmin = waveform.min()
 #        vmax = waveform.max()
 #        length = len(waveform)
-#        logging.debug(f'{channel_name}: V({vmin*1000:6.3f}, {vmax*1000:6.3f}) {length}')
+#        logger.debug(f'{channel_name}: V({vmin*1000:6.3f}, {vmax*1000:6.3f}) {length}')
         if channel_name in self.awg_channels:
             awg_name = self.awg_channels[channel_name].awg_name
         elif channel_name in self.marker_channels:
@@ -149,7 +150,7 @@ class QsUploader:
         awg_name = marker_channel.module_name
         awg = self.AWGs[awg_name]
         awg.load_marker_table(table)
-        logging.debug(f'marker for {channel_name} loaded in {(time.perf_counter()-start)*1000:4.2f} ms')
+        logger.debug(f'marker for {channel_name} loaded in {(time.perf_counter()-start)*1000:4.2f} ms')
 
     def __get_job(self, seq_id, index):
         """
@@ -164,7 +165,7 @@ class QsUploader:
             if job.seq_id == seq_id and job.index == index and not job.released:
                 return job
 
-        logging.error(f'Job not found for index {index} of seq {seq_id}')
+        logger.error(f'Job not found for index {index} of seq {seq_id}')
         raise ValueError(f'Sequence with id {seq_id}, index {index} not placed for upload .. . Always make sure to first upload your segment and then do the playback.')
 
 
@@ -279,9 +280,9 @@ class QsUploader:
                     else:
                         schedule.append(AwgInstruction(i, entry.time_after, wave_number=entry.waveform_index))
                 t3 = time.perf_counter()
-                logging.debug(f'{awg_sequencer.channel_name} create waves:{(t2-t1)*1000:6.3f}, seq:{(t3-t2)*1000:6.3f} ms')
+                logger.debug(f'{awg_sequencer.channel_name} create waves:{(t2-t1)*1000:6.3f}, seq:{(t3-t2)*1000:6.3f} ms')
             seq.load_schedule(schedule)
-        logging.debug(f'loaded awg sequences in {(time.perf_counter() - start)*1000:6.3f} ms')
+        logger.debug(f'loaded awg sequences in {(time.perf_counter() - start)*1000:6.3f} ms')
 
         start = time.perf_counter()
         for dig_channel in self.digitizer_channels.values():
@@ -303,7 +304,7 @@ class QsUploader:
                                                              threshold=entry.threshold))
                     seq.load_schedule(schedule)
 
-        logging.debug(f'loaded dig sequences in {(time.perf_counter() - start)*1000:6.3f} ms')
+        logger.debug(f'loaded dig sequences in {(time.perf_counter() - start)*1000:6.3f} ms')
 
         # start hvi (start function loads schedule if not yet loaded)
         acquire_triggers = {f'dig_trigger_{i+1}':t for i,t in enumerate(job.digitizer_triggers)}
@@ -391,7 +392,7 @@ class Job(object):
 
         self.channel_queues = dict()
         self.hw_schedule = None
-        logging.debug(f'new job {seq_id}-{index}')
+        logger.debug(f'new job {seq_id}-{index}')
 
 
     def add_hw_schedule(self, hw_schedule, schedule_params):
@@ -416,11 +417,11 @@ class Job(object):
 
     def release(self):
         if self.released:
-            logging.warning(f'job {self.seq_id}-{self.index} already released')
+            logger.warning(f'job {self.seq_id}-{self.index} already released')
             return
 
         self.upload_info = None
-        logging.debug(f'release job {self.seq_id}-{self.index}')
+        logger.debug(f'release job {self.seq_id}-{self.index}')
         self.released = True
 
         for channel_name, queue in self.channel_queues.items():
@@ -433,7 +434,7 @@ class Job(object):
 
     def __del__(self):
         if not self.released:
-            logging.warning(f'Job {self.seq_id}-{self.index} was not released. '
+            logger.warning(f'Job {self.seq_id}-{self.index} was not released. '
                             'Automatic release in destructor.')
             self.release()
 
@@ -573,7 +574,7 @@ class UploadAggregator:
                     else:
                         seg_ch = seg[channel_name]
                     channel_info.integral += seg_ch.integrate(job.index, sample_rate)
-                    logging.debug(f'Integral seg:{iseg} {channel_name} integral:{channel_info.integral}')
+                    logger.debug(f'Integral seg:{iseg} {channel_name} integral:{channel_info.integral}')
 
 
     def _generate_sections(self, job):
@@ -659,7 +660,7 @@ class UploadAggregator:
 
         # add DC compensation
         compensation_time = self.get_max_compensation_time()
-        logging.debug(f'DC compensation time: {compensation_time*1e9} ns')
+        logger.debug(f'DC compensation time: {compensation_time*1e9} ns')
         compensation_npt = int(np.ceil(compensation_time * section.sample_rate * 1e9))
         if compensation_npt > 50_000:
             # more than 50_000 samples? Use new segment with lower sample rate for compensation
@@ -677,7 +678,7 @@ class UploadAggregator:
             sections.append(section)
             # calculate npt
             compensation_npt = int(np.ceil(compensation_time * section.sample_rate * 1e9))
-            logging.info(f'Added new segment for DC compensation: {int(compensation_time*1e9)} ns, '
+            logger.info(f'Added new segment for DC compensation: {int(compensation_time*1e9)} ns, '
                          f'sample_rate: {sr/1e6} MHz, {compensation_npt} Sa')
 
         job.upload_info.dc_compensation_duration = compensation_npt/section.sample_rate
@@ -688,13 +689,13 @@ class UploadAggregator:
         section.align(extend=True)
         job.playback_time = section.t_end - sections[0].t_start
         job.n_waveforms = len(sections)
-        logging.debug(f'Playback time: {job.playback_time} ns')
+        logger.debug(f'Playback time: {job.playback_time} ns')
 
         if UploadAggregator.verbose:
             for segment in segments:
-                logging.info(f'segment: {segment}')
+                logger.info(f'segment: {segment}')
             for section in sections:
-                logging.info(f'section: {section}')
+                logger.info(f'section: {section}')
 
 
     def _generate_upload_wvf(self, job, awg_upload_func):
@@ -736,7 +737,7 @@ class UploadAggregator:
                 n_delay = iround(channel_info.delay_ns * sample_rate)
 
                 if isinstance(seg, conditional_segment):
-                    logging.debug(f'conditional for {channel_name}')
+                    logger.debug(f'conditional for {channel_name}')
                     seg_ch = get_conditional_channel(seg, channel_name)
                 else:
                     seg_ch = seg[channel_name]
@@ -746,15 +747,15 @@ class UploadAggregator:
                 #print(f'start: {channel_name}.{iseg}: {ref_channel_states.start_time}')
                 wvf = seg_ch.get_segment(job.index, sample_rate*1e9, ref_channel_states)
                 duration = time.perf_counter() - start
-                logging.debug(f'generated [{job.index}]{iseg}:{channel_name} {len(wvf)} Sa, in {duration*1000:6.3f} ms')
+                logger.debug(f'generated [{job.index}]{iseg}:{channel_name} {len(wvf)} Sa, in {duration*1000:6.3f} ms')
 
                 if len(wvf) != seg_render.npt:
-                    logging.warning(f'waveform {iseg}:{channel_name} {len(wvf)} Sa <> sequence length {seg_render.npt}')
+                    logger.warning(f'waveform {iseg}:{channel_name} {len(wvf)} Sa <> sequence length {seg_render.npt}')
 
                 i_start = 0
                 if seg_render.start_section:
                     if section != seg_render.start_section:
-                        logging.error(f'OOPS section mismatch {iseg}, {channel_name}')
+                        logger.error(f'OOPS section mismatch {iseg}, {channel_name}')
 
                     # add n_start_transition - n_delay to start_section
 #                    n_delay_welding = iround(channel_info.delay_ns * section.sample_rate)
@@ -807,7 +808,7 @@ class UploadAggregator:
 
                 else:
                     if section != seg_render.section:
-                        logging.error(f'OOPS-2 section mismatch {iseg}, {channel_name}')
+                        logger.error(f'OOPS-2 section mismatch {iseg}, {channel_name}')
                     offset = seg_render.offset + n_delay
                     buffer[offset+i_start:offset + len(wvf)] = wvf[i_start:]
 
@@ -821,7 +822,7 @@ class UploadAggregator:
                                      section.sample_rate, awg_upload_func)
                     section = sections[-1]
                     buffer = np.zeros(section.npt)
-                    logging.info(f'DC compensation section with {section.npt} Sa')
+                    logger.info(f'DC compensation section with {section.npt} Sa')
 
                 compensation_npt = iround(job.upload_info.dc_compensation_duration * section.sample_rate)
 
@@ -829,7 +830,7 @@ class UploadAggregator:
                     compensation_voltage = -channel_info.integral * section.sample_rate / compensation_npt * 1e9
                     job.upload_info.dc_compensation_voltages[channel_name] = compensation_voltage
                     buffer[-(compensation_npt+1):-1] = compensation_voltage
-                    logging.debug(f'DC compensation {channel_name}: {compensation_voltage:6.1f} mV {compensation_npt} Sa')
+                    logger.debug(f'DC compensation {channel_name}: {compensation_voltage:6.1f} mV {compensation_npt} Sa')
                 else:
                     job.upload_info.dc_compensation_voltages[channel_name] = 0
 
@@ -840,7 +841,7 @@ class UploadAggregator:
 
     def _render_markers(self, job, awg_upload_func):
         for channel_name, marker_channel in self.marker_channels.items():
-            logging.debug(f'Marker: {channel_name} ({marker_channel.amplitude} mV, {marker_channel.delay:+2.0f} ns)')
+            logger.debug(f'Marker: {channel_name} ({marker_channel.amplitude} mV, {marker_channel.delay:+2.0f} ns)')
             start_stop = []
             if channel_name in self.rf_marker_pulses:
                 offset = marker_channel.delay
@@ -852,7 +853,7 @@ class UploadAggregator:
                 offset = seg_render.t_start + marker_channel.delay
 
                 if isinstance(seg, conditional_segment):
-                    logging.debug(f'conditional for {channel_name}')
+                    logger.debug(f'conditional for {channel_name}')
                     seg_ch = get_conditional_channel(seg, channel_name)
                 else:
                     seg_ch = seg[channel_name]
@@ -885,19 +886,19 @@ class UploadAggregator:
         for on_off in m:
             s += on_off[1]
             if s < 0:
-                logging.error(f'Marker error {marker_channel.name} {on_off}')
+                logger.error(f'Marker error {marker_channel.name} {on_off}')
             if s == 1 and on_off[1] == 1:
                 t_on = on_off[0]
             if s == 0 and on_off[1] == -1:
                 t_off = on_off[0]
-                logging.debug(f'Marker: {t_on} - {t_off}')
+                logger.debug(f'Marker: {t_on} - {t_off}')
                 # search start section
                 while t_on >= sections[i_section].t_end:
                     i_section += 1
                 section = sections[i_section]
                 pt_on = int((t_on - section.t_start) * section.sample_rate)
                 if pt_on < 0:
-                    logging.info(f'Warning: Marker setup before waveform; aligning with start')
+                    logger.info(f'Warning: Marker setup before waveform; aligning with start')
                     pt_on = 0
                 if t_off < section.t_end:
                     pt_off = int((t_off - section.t_start) * section.sample_rate)
@@ -925,13 +926,13 @@ class UploadAggregator:
         for on_off in m:
             s += on_off[1]
             if s < 0:
-                logging.error(f'Marker error {marker_channel.name} {on_off}')
+                logger.error(f'Marker error {marker_channel.name} {on_off}')
             if s == 1 and on_off[1] == 1:
                 t_on = int(on_off[0])
             if s == 0 and on_off[1] == -1:
                 t_off = int(on_off[0])
                 if UploadAggregator.verbose:
-                    logging.debug(f'Marker: {t_on} - {t_off}')
+                    logger.debug(f'Marker: {t_on} - {t_off}')
                 table.append((t_on + offset, t_off + offset))
 
     def _upload_wvf(self, job, channel_name, waveform, amplitude, attenuation, sample_rate, awg_upload_func):
@@ -951,7 +952,7 @@ class UploadAggregator:
 
         for channel_name, qubit_channel in self.qubit_channels.items():
             if channel_name not in self.sequencer_channels:
-                logging.warning(f'QS driver (M3202A_QS) not loaded for qubit channel {channel_name}')
+                logger.warning(f'QS driver (M3202A_QS) not loaded for qubit channel {channel_name}')
                 continue
             start = time.perf_counter()
             delays = []
@@ -984,7 +985,7 @@ class UploadAggregator:
                         else:
                             sequence.shift_phase(t_pulse, e.phase_shift)
                 else:
-                    logging.debug(f'conditional for {channel_name}:{iseg} start:{seg_render.t_start}')
+                    logger.debug(f'conditional for {channel_name}:{iseg} start:{seg_render.t_start}')
                     cond_ch = get_conditional_channel(seg, channel_name, sequenced=True, index=job.index)
                     qs_cond = self.conditional_segments[iseg]
 
@@ -999,7 +1000,7 @@ class UploadAggregator:
                                                     condition_register=3)
             sequence.close()
             duration = time.perf_counter() - start
-            logging.debug(f'generated iq sequence {channel_name} {duration*1000:6.3f} ms')
+            logger.debug(f'generated iq sequence {channel_name} {duration*1000:6.3f} ms')
 
 #    def _generate_sequencer_baseband_upload(self, job):
 # TODO @@@ baseband pulses
@@ -1041,7 +1042,7 @@ class UploadAggregator:
             t_end = None
             for iseg, (seg, seg_render) in enumerate(zip(job.sequence, self.segments)):
                 if isinstance(seg, conditional_segment):
-                    logging.debug(f'conditional for {channel_name}')
+                    logger.debug(f'conditional for {channel_name}')
                     seg_ch = get_conditional_channel(seg, channel_name)
                 else:
                     seg_ch = seg[channel_name]
@@ -1078,7 +1079,7 @@ class UploadAggregator:
                 triggers.append([nr for module_name, nr in all_channels if module_name == name])
 
         job.digitizer_trigger_channels = digitizer_trigger_channels
-        logging.info(f'digitizer triggers: {job.digitizer_triggers}')
+        logger.info(f'digitizer triggers: {job.digitizer_triggers}')
 
     def _generate_digitizer_sequences(self, job):
 
@@ -1095,7 +1096,7 @@ class UploadAggregator:
                     pxi_triggers[acq] = pxi
                     pxi += 1
 
-        logging.debug(f'PXI triggers: {pxi_triggers}')
+        logger.debug(f'PXI triggers: {pxi_triggers}')
 
         segments = self.segments
         for channel_name, channel in self.digitizer_channels.items():
@@ -1113,7 +1114,7 @@ class UploadAggregator:
             for iseg, (seg, seg_render) in enumerate(zip(job.sequence, segments)):
                 offset = channel.delay + int(self.max_pre_start_ns)
                 if isinstance(seg, conditional_segment):
-                    logging.debug(f'conditional for {channel_name}')
+                    logger.debug(f'conditional for {channel_name}')
                     # TODO @@@ lookup acquisitions and set pxi trigger.
                     seg_ch = get_conditional_channel(seg, channel_name)
                 else:
@@ -1138,7 +1139,7 @@ class UploadAggregator:
                                      threshold=acquisition.threshold,
                                      pxi_trigger=pxi_trigger)
 
-                    logging.debug(f'Acq: {acquisition.ref}: {pxi_trigger}')
+                    logger.debug(f'Acq: {acquisition.ref}: {pxi_trigger}')
                     t_end = t+t_measure
                     if rf_source is not None and rf_source.mode != 'continuous':
                         rf_marker_pulses.append(RfMarkerPulse(t, t_end))
@@ -1223,7 +1224,7 @@ class UploadAggregator:
             compensation_factor = 1 / (sample_rate * 1e9 * channel_info.bias_T_RC_time)
             compensation = np.cumsum(buffer) * compensation_factor + bias_T_compensation_mV
             bias_T_compensation_mV = compensation[-1]
-            logging.info(f'bias-T compensation  min:{np.min(compensation):5.1f} max:{np.max(compensation):5.1f} mV')
+            logger.info(f'bias-T compensation  min:{np.min(compensation):5.1f} max:{np.max(compensation):5.1f} mV')
             buffer += compensation
 
         return bias_T_compensation_mV